feat(order): add validation process for orders
