--- src/main/kotlin/ecommerce/dto/OrderRequest.kt
@@ -1,6 +1,0 @@
-package ecommerce.dto
-
-class OrderRequest(
-    val currency: String,
-    val paymentMethod: String,
-)
