--- src/main/kotlin/ecommerce/dto/OrderResponse.kt
@@ -1,7 +1,0 @@
-package ecommerce.dto
-
-class OrderResponse(
-    val orderId: Long,
-    val orderItems: List<OrderItemResponse>,
-    val currency: String,
-)
